refactor: unify api endpoints to v1 and implement status-based post filtering

+++ src/main/java/com/abdelrahman/blogplatorm/dtos/responses/CategoryResponseDto.java
@@ -12,5 +12,5 @@ public class CategoryResponseDto {
 	
 	private String name;
 	
-	private Integer postCount;
+	private long postCount;
 }
